requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.9.0

# Google AI & Cloud
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.14.0

# Agent Orchestration
langgraph==0.0.26
langchain==0.1.0
langchain-google-genai==0.0.5

# Machine Learning
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
tensorflow==2.15.0
scikit-learn==1.3.2
xgboost==2.0.3

# Data Processing
pandas==2.1.4
numpy==1.26.2
geopandas==0.14.1
shapely==2.0.2

# Visualization
plotly==5.18.0
folium==0.15.1
matplotlib==3.8.2
seaborn==0.13.0

# Communication APIs
twilio==8.11.0
requests==2.31.0

# Web Framework (for dashboard)
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
flask==3.0.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
pillow==10.1.0
opencv-python==4.8.1.78
